+++ packages/vite/types/dynamicImportVars.d.ts
@@ -1,1 +1,0 @@
-export type { RollupDynamicImportVarsOptions } from '../dist/node'